requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
fastapi>=0.110.0
uvicorn>=0.27.0
httpx>=0.27.0
